move tests to test_setitem.py
